Fix OD6211 Adjust user mnemonics file name and locations
Summary:
Fixes OD-6211
- look in both data root and current survey directory for user mnemonics file
- use GetInterpreterName function to make file name
- fix issue with writing linestyle color to file

Test Plan: Tested on Linux

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com/D17833

+++ src/uiTools/uimnemonicprops.cc
@@ -111,9 +111,9 @@ BufferString uiMnemonicProperties::toString() const
 	.add( getUOMStr() )
 	.add( OD::LineStyle::toString(ls.type_) )
 	.add( ls.width_ )
-	.add( lcol.r() )
-	.add( lcol.g() )
-	.add( lcol.b() );
+	.add( (int)lcol.r() )
+	.add( (int)lcol.g() )
+	.add( (int)lcol.b() );
 
     res = fms;
     return res;
